requirements.txt
fastapi
uvicorn
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
starlette==0.47.3

pdfminer-six==20250506
pymupdf==1.26.4
pypdf==6.0.0
fitz

torch==2.2.2
transformers==4.55.4
tokenizers==0.21.4
huggingface-hub==0.34.4
safetensors==0.6.2

requests==2.32.5
tqdm==4.67.1
regex==2025.7.34
numpy==1.26.4
